requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
google-generativeai
pydantic
